requirements.txt
flask
flask-sqlalchemy
flask-socketio
Flask-Caching
pycryptodome
requests